Extract memeolist schema to reuse it in multiple configs
